+++ modeldict/base.py
@@ -183,6 +183,9 @@ class CachedDict(object):
             if local_cache_is_invalid or local_cache_is_invalid is None:
                 self._local_cache = self.remote_cache.get(self.remote_cache_key)
 
+            # We last checked for remote changes just now
+            self._last_checked_for_remote_changes = now
+
             # No matter what, we've updated from remote, so mark ourselves as
             # such so that we won't expire until the next timeout
             self._local_last_updated = now
@@ -190,9 +193,6 @@ class CachedDict(object):
         # Update from cache if local_cache is still empty
         if self._local_cache is None:
             self._update_cache_data()
-
-        # No matter what happened, we last checked for remote changes just now
-        self._last_checked_for_remote_changes = now
 
         return self._local_cache